마이페이지 마크 라이크 개수 응답값 추가

+++ application/g-app-application/src/main/java/com/nelly/application/dto/response/GetMyPageResponse.java
@@ -32,8 +32,8 @@ public class GetMyPageResponse {
     private String profileText;
     private String status;
     private int followerCount;
-    private int likeCount;
-    private int markCount;
+    private long likeCount;
+    private long markCount;
     private int contentsCount;
     private int contentMarkCount;
     private boolean followed;
